discovery: fix+consolidate Zookeeper discoveries
